refactor: 입력 관련 로직 domain 패키지로 이동
