+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,9 +1,8 @@
-package lotto;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import lotto.domain.Lotto;
 import lotto.validation.LottoNumbersValidator;
 import lotto.view.input.InvalidInputException;
 import org.junit.jupiter.api.DisplayName;
