+++ src/test/java/lotto/domain/PurchasedPriceTest.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import lotto.validation.LottoPurchaseValidator;
 import org.junit.jupiter.api.DisplayName;
@@ -9,6 +10,20 @@ import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 public class PurchasedPriceTest {
+    @ParameterizedTest
+    @ValueSource(longs = {1000, 10000, 100000})
+    @DisplayName("올바른 구매 금액 입력")
+    void 올바른_구매_금액_입력(long input) {
+        // given
+        long expectedCount = input / 1000;
+
+        // when
+        LottoCount lottoCount = LottoCount.calculatePurchaseCount(input);
+
+        // then
+        assertEquals(expectedCount, lottoCount.getCount());
+    }
+
     @ParameterizedTest
     @NullAndEmptySource
     @DisplayName("구매 금액을 입력하지 않았다면 예외가 발생한다.")
